# docs/backend-contracts/contract-history/1.3.0-contract-914084874afda2481813a041da4cc01249aa9ea557d9a8bf29baeed4f10e0dc9/openapi.snapshot.yaml
openapi: 3.1.0
info:
  title: BNBU Sports Unified API
  version: 1.3.0-contract
  summary: Unified contract for the sports check-in platform
  description: |
    Stage 21B contract. It does not prove that every registered capability is
    business-enabled or production-ready.
    Business fields come from 02-data-dictionary.md, state transitions from
    03-state-machines.md, rules from 04-business-rules.md, and authorization
    from 05-permission-matrix.md. Foundation authentication/session persistence
    is approved; CourseInvite/Join Capability are contract-and-persistence-design
    only until the Enrollment Gate, while every Export operation is fail-closed
    and V1 persists no ExportJob. Of the 30 client-support operations added in
    Stage 21, authentication, exemption, release-policy, notification,
    push-device, preference, help, and feedback capabilities are approved for
    real local implementation, while unresolved capabilities remain explicit
    routed denials. GPS has persistence and
    application-layer foundations, but all six location HTTP operations remain
    denied. None of these statements proves Staging, iOS-binary, push-delivery,
    GPS-production, or Production-Gate readiness.
servers:
  - url: /api/v1
    description: Version 1 relative base path
security:
  - bearerAuth: []
...
paths:
...
  /auth/password-login:
    post:
      tags: [Authentication]
      operationId: passwordLogin
      x-access-policy:
        policyId: AUTH-PASSWORD-LOGIN
        authentication: PUBLIC
        allowedRoles: []
        organizationScope: NONE
        resourceScope: NONE
        resourceResolver: NONE
        defaultDeny: true
      summary: Authenticate with an approved password credential
      description: Foundation password login for pre-provisioned TEACHER/ADMIN accounts; TTL values come from environment configuration and arbitrary-password or role-guessing login is forbidden.
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
            example:
              account: teacher@example.edu
              password: example-only-not-a-real-password
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
              example:
                data:
                  sessionId: ses_auth_01JABC123
                  accessToken: eyJhbGciOiJIUzI1NiJ9.example
                  refreshToken: opaque-refresh-token
                  tokenType: Bearer
                  accessTokenExpiresAt: '2026-08-02T02:20:00Z'
                  refreshTokenExpiresAt: '2026-08-09T02:05:00Z'
                  user:
                    id: usr_01JABC123
                    organizationId: org_bnbu
                    role: TEACHER
                    status: ACTIVE
                    primaryEmailMasked: t***@example.edu
                    primaryPhoneMasked: null
                    emailVerified: true
                    phoneVerified: false
                    version: 4
                meta:
                  requestId: req_01JABC123
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      x-access-policy:
        policyId: AUTH-REFRESH
        authentication: PUBLIC
        allowedRoles: []
        organizationScope: NONE
        resourceScope: SESSION
        resourceResolver: REFRESH_TOKEN
        defaultDeny: true
      summary: Rotate a refresh token and issue a new access token
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logoutSession
      x-access-policy:
        policyId: AUTH-LOGOUT
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: SESSION
        resourceResolver: AUTHENTICATED_SESSION
        defaultDeny: true
      summary: Revoke the current or selected owned session
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /class-sections:
    get:
      tags: [Class Sections]
      operationId: listClassSections
      x-access-policy:
        policyId: CLASS-SECTION-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: CLASS_SECTION_LIST_SCOPE
        defaultDeny: true
      summary: List authorized class sections
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
        - name: courseId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: semesterId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/ClassSectionListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    post:
      tags: [Class Sections]
      operationId: createClassSection
      x-access-policy:
        policyId: CLASS-SECTION-CREATE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: CLASS_SECTION_FROM_REQUEST
        defaultDeny: true
      summary: Create a class section owned by the authenticated responsible teacher
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassSectionRequest'
      responses:
        '201':
          $ref: '#/components/responses/ClassSectionSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /class-sections/{classSectionId}:
    parameters:
      - $ref: '#/components/parameters/ClassSectionId'
    get:
      tags: [Class Sections]
      operationId: getClassSection
      x-access-policy:
        policyId: CLASS-SECTION-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: CLASS_SECTION_FROM_PATH
        defaultDeny: true
      summary: Get an authorized class section projection
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/ClassSectionSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Class Sections]
      operationId: updateClassSection
      x-access-policy:
        policyId: CLASS-SECTION-UPDATE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: CLASS_SECTION_FROM_PATH
        defaultDeny: true
      summary: Update role-authorized class section fields
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassSectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/ClassSectionSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /class-sections/{classSectionId}/close:
    post:
      tags: [Class Sections]
      operationId: closeClassSection
      x-access-policy:
        policyId: CLASS-SECTION-CLOSE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: CLASS_SECTION_FROM_PATH
        defaultDeny: true
      summary: Close a class section owned by the responsible teacher
      parameters:
        - $ref: '#/components/parameters/ClassSectionId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          $ref: '#/components/responses/ClassSectionSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token carried only in the Authorization header.
    JoinCapability:
      type: apiKey
      in: header
      name: X-Join-Capability
      description: Short-lived one-time capability bound to one ClassSection. The plaintext value is never logged or persisted; no ordinary Access Token is accepted by the Join operation.
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; the server returns the accepted/generated value in response header and meta.requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Reuse for retries of the same normalized request; reuse with different content returns a conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to caller scope, filters, and sort.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated endpoint whitelist; prefix a field with '-' for descending order.
      schema:
        type: string
        maxLength: 200
    Search:
      name: q
      in: query
      required: false
      description: Search only endpoint-declared fields; never expands authorization scope.
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
    ClassSectionId:
      name: classSectionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
...
  responses:
    EmptySuccess:
      description: Successful action with no resource body
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EmptyEnvelope' }
...
    AuthSuccess:
      description: Authenticated or refreshed session
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthEnvelope' }
...
    ClassSectionSuccess:
      description: Class section projection
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ClassSectionEnvelope' }
    ClassSectionListSuccess:
      description: Paged class section projections
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ClassSectionListEnvelope' }
...
    BadRequest:
      $ref: '#/components/responses/Error400'
    Unauthorized:
      $ref: '#/components/responses/Error401'
    Forbidden:
      $ref: '#/components/responses/Error403'
    NotFound:
      $ref: '#/components/responses/Error404'
...
    Conflict:
      $ref: '#/components/responses/Error409'
    UnprocessableEntity:
      $ref: '#/components/responses/Error422'
    TooManyRequests:
      $ref: '#/components/responses/Error429'
    InternalError:
      $ref: '#/components/responses/Error500'
...
    Error400:
      description: Bad request
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error401:
      description: Unauthorized
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error403:
      description: Forbidden
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error404:
      description: Not found
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error409:
      description: Conflict
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error422:
      description: Unprocessable entity
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error429:
      description: Too many requests
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
        Retry-After:
          description: Seconds until a retry may be attempted.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error500:
      description: Internal server error
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  headers:
    RequestIdResponseHeader:
      description: Server-accepted/generated request ID; equals meta.requestId on success and requestId on error.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._:-]+$'
      description: Opaque internal ID; clients must not parse prefixes or ordering.
...
    PositiveVersion:
      type: integer
      format: int64
      minimum: 1
...
    UserRole:
      type: string
      enum: [STUDENT, TEACHER, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, LOCKED, DISABLED]
...
    ClassSectionStatus:
      type: string
      enum: [UPCOMING, ACTIVE, CLOSED, ARCHIVED]
    CheckInWindowMode:
      type: string
      enum: [AVAILABLE, UNAVAILABLE]
...
    PaginationMeta:
      type: object
      additionalProperties: false
      properties:
        nextCursor:
          type: [string, 'null']
          maxLength: 2048
        hasMore:
          type: boolean
        limit:
          type: integer
          minimum: 1
          maximum: 100
      required: [nextCursor, hasMore, limit]
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        i18nKey:
          type: string
          minLength: 1
          maxLength: 300
        params:
          type: object
          additionalProperties: true
      required: [field, code, i18nKey, params]
...
    ErrorDetails:
      type: object
      additionalProperties: false
      properties:
        fieldErrors:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/FieldError' }
        resourceType:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        resourceId:
          type: string
          minLength: 1
          maxLength: 64
        currentState:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        allowedActions:
          type: array
          maxItems: 100
          uniqueItems: true
          items: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        expectedVersion:
          type: integer
          minimum: 0
        actualVersion:
          type: integer
          minimum: 0
        retryAfterSeconds:
          type: integer
          minimum: 0
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 128
        itemErrors:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/BatchItemError' }
        migrationReference:
          type: string
          minLength: 1
          maxLength: 256
    BatchItemError:
      type: object
      additionalProperties: false
      properties:
        itemId:
          type: string
          minLength: 1
          maxLength: 128
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
        details: { $ref: '#/components/schemas/ErrorDetails' }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        timestamp:
          type: string
          format: date-time
      required: [itemId, code, message, details, requestId, timestamp]
...
    SuccessMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 64
      required: [requestId]
    PagedMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
      required: [requestId, pagination]
    ErrorEnvelope:
      type: object
      additionalProperties: false
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
        details: { $ref: '#/components/schemas/ErrorDetails' }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        timestamp:
          type: string
          format: date-time
      required: [code, message, details, requestId, timestamp]
      example:
        code: CONFLICT_VERSION_MISMATCH
        message: The resource changed. Refresh and try again.
        details:
          resourceType: EXERCISE_RECORD
          resourceId: rec_01JABC123
          expectedVersion: 3
          actualVersion: 4
        requestId: req_01JABC123
        timestamp: '2026-08-02T12:00:00Z'
    EmptyEnvelope:
      type: object
      additionalProperties: false
      properties:
        data: { type: 'null' }
        meta: { $ref: '#/components/schemas/SuccessMeta' }
      required: [data, meta]
...
    User:
      type: object
      additionalProperties: false
      description: Minimal authorized projection; authentication internals and plaintext contact methods are never returned.
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        primaryEmailMasked: { type: [string, 'null'], maxLength: 254, readOnly: true, description: Masked projection only; plaintext email is never returned. }
        primaryPhoneMasked: { type: [string, 'null'], maxLength: 32, readOnly: true, description: Masked projection only; plaintext phone is never returned. }
        emailVerified: { type: boolean, readOnly: true, description: Derived from the current contact verification fact; the verification timestamp is not exposed. }
        phoneVerified: { type: boolean, readOnly: true, description: Derived from the current contact verification fact; the verification timestamp is not exposed. }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, role, status, primaryEmailMasked, primaryPhoneMasked, emailVerified, phoneVerified, version]
...
    ClassSection:
      type: object
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        courseId: { $ref: '#/components/schemas/OpaqueId' }
        semesterId: { $ref: '#/components/schemas/OpaqueId' }
        teacherId: { $ref: '#/components/schemas/OpaqueId' }
        classCode: { type: string, minLength: 1, maxLength: 64 }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        status: { $ref: '#/components/schemas/ClassSectionStatus' }
        isEnrollmentOpen: { type: boolean }
        checkInWindowMode: { $ref: '#/components/schemas/CheckInWindowMode' }
        checkInStartDate: { type: [string, 'null'], format: date }
        checkInEndDate: { type: [string, 'null'], format: date }
        dailyStartTime: { type: [string, 'null'], format: time }
        dailyEndTime: { type: [string, 'null'], format: time }
        submissionDeadlineAt: { type: [string, 'null'], format: date-time }
        excludedDates:
          type: array
          uniqueItems: true
          items: { type: string, format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, courseId, semesterId, teacherId, classCode, displayName, status, isEnrollmentOpen, checkInWindowMode, excludedDates, createdAt, updatedAt, version]
...
    AuthSession:
      type: object
      additionalProperties: false
      required: [sessionId, accessToken, refreshToken, tokenType, accessTokenExpiresAt, refreshTokenExpiresAt, user]
      properties:
        sessionId: { oneOf: [{ $ref: '#/components/schemas/OpaqueId' }, { type: 'null' }] }
        enrollmentId: { oneOf: [{ $ref: '#/components/schemas/OpaqueId' }, { type: 'null' }] }
        accessToken: { type: string, minLength: 1 }
        refreshToken: { type: string, minLength: 1 }
        tokenType: { type: string, const: Bearer }
        accessTokenExpiresAt: { type: string, format: date-time }
        refreshTokenExpiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
      description: Plaintext tokens are returned only at issuance/rotation. The server persists AuthSession plus irreversible RefreshToken digests; all TTLs come from environment configuration. Production browser refresh transport remains a separate security decision.
...
    PasswordLoginRequest:
      type: object
      additionalProperties: false
      required: [account, password]
      properties:
        account: { type: string, minLength: 1, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1, writeOnly: true }
    LogoutRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1, writeOnly: true }
...
    CreateClassSectionRequest:
      type: object
      additionalProperties: false
      required: [courseId, semesterId, classCode, displayName]
      properties:
        courseId: { $ref: '#/components/schemas/OpaqueId' }
        semesterId: { $ref: '#/components/schemas/OpaqueId' }
        classCode: { type: string, minLength: 1, maxLength: 64 }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        isEnrollmentOpen: { type: boolean, default: false }
      description: The responsible teacherId and organizationId are always derived from the authenticated TEACHER principal; client-supplied identity fields are rejected.
    UpdateClassSectionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 200 }
        isEnrollmentOpen: { type: boolean }
        checkInWindowMode: { $ref: '#/components/schemas/CheckInWindowMode' }
        checkInStartDate: { type: [string, 'null'], format: date }
        checkInEndDate: { type: [string, 'null'], format: date }
        dailyStartTime: { type: [string, 'null'], format: time }
        dailyEndTime: { type: [string, 'null'], format: time }
        submissionDeadlineAt: { type: [string, 'null'], format: date-time }
        excludedDates: { type: array, uniqueItems: true, items: { type: string, format: date } }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
...
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [reason, expectedVersion]
      properties:
        reason: { type: string, minLength: 1, maxLength: 1000 }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
...
    AuthEnvelope: { $ref: '#/components/schemas/AuthResponse' }
...
    ClassSectionEnvelope: { $ref: '#/components/schemas/ClassSectionResponse' }
    ClassSectionListEnvelope: { $ref: '#/components/schemas/ClassSectionListResponse' }
...
    AuthResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/AuthSession' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
...
    ClassSectionResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/ClassSection' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
    ClassSectionListResponse:
      type: object
      required: [data, meta]
      properties: { data: { type: array, items: { $ref: '#/components/schemas/ClassSection' } }, meta: { $ref: '#/components/schemas/PagedMeta' } }
...